% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanic.R
\docType{package}
\name{sanic}
\alias{sanic}
\title{Solving Ax = b Nimbly in C++}
\description{
Fast routines for solving large systems of linear equations in
R. Makes Eigen's Cholesky-, LU-, QR-, and iterative (Conjugate Gradient,
BiCGSTAB) solvers for both dense and sparse problems available.
}
