% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_iterative.R
\name{solve_cg}
\alias{solve_cg}
\title{Solve a System of Equations using Iterative Methods}
\usage{
solve_cg(
  a,
  b,
  x0,
  type = c("BiCGSTAB", "LSCG", "CG"),
  tol,
  iter,
  verbose = FALSE
)
}
\arguments{
\item{a}{Square numeric matrix with the coefficients of the linear system.
Both dense and sparse matrices are supported (see \link{sparsify}).}

\item{b}{Numeric vector or matrix at the right-hand side of the linear
system. If missing, 'b' is set to an identity matrix and 'a' is
inverted.}

\item{x0}{Initial guess}

\item{type}{Whether to use the BiCGSTAB, least squares CG or CG solver}

\item{tol}{Tolerance}

\item{iter}{Iterations}

\item{verbose}{Whether to print iterations and tolerance}
}
\value{
Solves for \eqn{x} and returns a numeric matrix with the results.
}
\description{
Solve a System of Equations using Iterative Methods
}
\examples{
solve_cg(matrix(rnorm(9), 3), rnorm(3))
}
