% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_direct.R
\name{solve_chol}
\alias{solve_chol}
\alias{solve_lu}
\alias{solve_qr}
\title{Solve a System of Equations Using Direct Methods}
\usage{
solve_chol(a, b, type = c("LDLT", "LLT"))

solve_lu(a, b)

solve_qr(a, b, pivot = TRUE)
}
\arguments{
\item{a}{Square numeric matrix with the coefficients of the linear system.
Both dense and sparse matrices are supported (see \link{sparsify}).}

\item{b}{Numeric vector or matrix at the right-hand side of the linear
system. If missing, 'b' is set to an identity matrix and 'a' is
inverted.}

\item{type}{Character scalar. Whether to decompose 'a' as LDLT or LLT.}

\item{pivot}{Logical scalar. Whether to perform column pivoting.}
}
\value{
Solves for \eqn{x} and returns a numeric matrix with the results.
}
\description{
Solve a System of Equations Using Direct Methods
}
\examples{
solve_chol(crossprod(matrix(rnorm(9), 3)), rnorm(3))
}
