% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_qr}
\alias{solve_qr}
\title{Solve using QR}
\usage{
solve_qr(A, b, pivot = FALSE)
}
\arguments{
\item{A}{lhs}

\item{b}{rhs}

\item{pivot}{Whether to pivot}
}
\description{
Solve using QR
}
\examples{
solve_qr(matrix(rnorm(9), 3), rnorm(3))
}
