% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sparsify}
\alias{sparsify}
\title{Transform a Matrix to Be Sparse.}
\usage{
sparsify(x)
}
\arguments{
\item{x}{Numeric matrix to transform to a sparse 'dgCMatrix'.}
}
\value{
Returns 'x' as \code{dgCMatrix}.
}
\description{
Concise function to transform dense to sparse matrices of class
\code{dgCMatrix} (see \link[Matrix]{sparseMatrix}).
}
\examples{
sparsify(matrix(rnorm(9L), 3L))
}
