% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_square}
\alias{is_square}
\title{Check Squarity}
\usage{
is_square(x)
}
\arguments{
\item{x}{Matrix or something else with dimensions.}
}
\value{
Returns a logical scalar indicating squarity.
}
\description{
Check Squarity
}
