% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_cg}
\alias{solve_cg}
\title{Solve using conjugate gradient}
\usage{
solve_cg(
  A,
  b,
  type = c("BiCGSTAB", "LSCG", "CG"),
  x0,
  tol,
  iter,
  verbose = FALSE
)
}
\arguments{
\item{A}{lhs}

\item{b}{rhs}

\item{type}{Whether to use the BiCGSTAB, least squares CG or CG solver}

\item{x0}{Initial guess}

\item{tol}{Tolerance}

\item{iter}{Iterations}

\item{verbose}{Whether to print iterations and tolerance}
}
\description{
Solve using conjugate gradient
}
\examples{
solve_cg(matrix(rnorm(9), 3), rnorm(3))
}
