% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_iterative.R
\name{solve_cg}
\alias{solve_cg}
\title{Solve a System of Equations using Iterative Methods}
\usage{
solve_cg(
  a,
  b,
  x0,
  type = c("BiCGSTAB", "LSCG", "CG"),
  tol,
  iter,
  verbose = FALSE
)
}
\arguments{
\item{a}{Square numeric matrix with the coefficients of the linear system.
Both dense and sparse matrices are supported (see \link{sparsify}).}

\item{b}{Numeric vector or matrix at the right-hand side of the linear
system. If missing, 'b' is set to an identity matrix and 'a' is
inverted.}

\item{x0}{Initial guess}

\item{type}{Whether to use the BiCGSTAB, least squares CG or CG solver}

\item{tol}{Tolerance}

\item{iter}{Iterations}

\item{verbose}{Whether to print iterations and tolerance}
}
\value{
Solves for \eqn{x} and returns a numeric matrix with the results.
}
\description{
Functions to access specific iterative solvers for systems of equations.
}
\examples{
# Solve via least squares or bi-conjugate gradient methods
A <- matrix(rnorm(9), nrow = 3, ncol = 3)
x <- rnorm(3)
b  <- A \%*\% x

x_bi <- solve_cg(A, b)
x_ls <- solve_cg(A, b, type = "LS")

# Solve via conjugate gradient for symmetric matrices
AA <- crossprod(A)
b <- AA \%*\% x
x_cg <- solve_cg(AA, b, type = "CG")

}
