% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_chol}
\alias{solve_chol}
\title{Solve using Cholesky}
\usage{
solve_chol(A, b, type = c("LDLT", "LLT"))
}
\arguments{
\item{A}{lhs}

\item{b}{rhs}

\item{type}{Whether to decompose to LDLT or LLT}
}
\description{
Solve using Cholesky
}
\examples{
solve_chol(matrix(rnorm(9), 3), rnorm(3))
}
